🚚 Move code into live.ditto.demo.kmp package
