+++ composeApp/src/androidInstrumentedTest/kotlin/live/ditto/demo/kmp/AndroidGreetingTest.kt
@@ -1,6 +1,5 @@
 package live.ditto.demo.kmp
 
-import Greeting
 import org.junit.Test
 import kotlin.test.assertTrue
