+++ composeApp/src/androidMain/kotlin/live/ditto/demo/kmp/Ditto.kt
@@ -1,14 +1,15 @@
-import Env.DITTO_APP_ID
-import Env.DITTO_OFFLINE_TOKEN
+package live.ditto.demo.kmp
+
+import Env
 import android.util.Log
 import live.ditto.Ditto
 import live.ditto.DittoConnectionRequestAuthorization.Allow
 import live.ditto.DittoIdentity.OfflinePlayground
-import live.ditto.DittoLogLevel
+import live.ditto.DittoLogLevel.INFO
 import live.ditto.DittoLogger
 import live.ditto.android.DefaultAndroidDittoDependencies
 import live.ditto.transports.DittoSyncPermissions
-import org.koin.java.KoinJavaComponent.getKoin
+import org.koin.java.KoinJavaComponent
 
 @Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
 actual open class Ditto actual constructor() {
@@ -16,13 +17,13 @@ actual open class Ditto actual constructor() {
         println("Ditto (android) init")
     }
 
-    private val dependencies = DefaultAndroidDittoDependencies(getKoin().get())
-    private val identity = OfflinePlayground(dependencies, DITTO_APP_ID)
+    private val dependencies = DefaultAndroidDittoDependencies(KoinJavaComponent.getKoin().get())
+    private val identity = OfflinePlayground(dependencies, Env.DITTO_APP_ID)
     private val ditto: Ditto =
         Ditto(dependencies, identity).also {
             DittoLogger.enabled = true
-            DittoLogger.minimumLogLevel = DittoLogLevel.INFO
-            it.setOfflineOnlyLicenseToken(DITTO_OFFLINE_TOKEN)
+            DittoLogger.minimumLogLevel = INFO
+            it.setOfflineOnlyLicenseToken(Env.DITTO_OFFLINE_TOKEN)
         }
 
     // Switch to ditto.sdkVersion in 4.7.2-rc.2
@@ -34,7 +35,7 @@ actual open class Ditto actual constructor() {
     actual open val presence = DittoPresence(ditto.presence)
 
     actual open fun startSync() {
-        val missingPermissions = DittoSyncPermissions(getKoin().get()).missingPermissions()
+        val missingPermissions = DittoSyncPermissions(KoinJavaComponent.getKoin().get()).missingPermissions()
         Log.d(TAG, "Missing permissions: $missingPermissions.")
 
         if (missingPermissions.isNotEmpty()) {
