+++ composeApp/src/androidMain/kotlin/live/ditto/demo/kmp/DittoApplication.kt
@@ -1,7 +1,6 @@
 package live.ditto.demo.kmp
 
 import android.app.Application
-import initKoin
 import org.koin.android.ext.koin.androidContext
 
 class DittoApplication : Application() {
