+++ composeApp/src/androidMain/kotlin/live/ditto/demo/kmp/MainActivity.kt
@@ -1,6 +1,5 @@
 package live.ditto.demo.kmp
 
-import App
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
