+++ composeApp/src/androidMain/kotlin/live/ditto/demo/kmp/Platform.android.kt
@@ -1,3 +1,5 @@
+package live.ditto.demo.kmp
+
 import android.os.Build
 
 class AndroidPlatform : Platform {
