+++ composeApp/src/androidUnitTest/kotlin/live/ditto/demo/kmp/AndroidGreetingTest.kt
@@ -1,6 +1,5 @@
 package live.ditto.demo.kmp
 
-import Greeting
 import org.junit.Assert
 import org.junit.Test
