+++ composeApp/src/commonMain/kotlin/live/ditto/demo/kmp/Ditto.kt
@@ -1,3 +1,5 @@
+package live.ditto.demo.kmp
+
 @Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
 expect open class Ditto() {
     open val version: String
