+++ composeApp/src/commonMain/kotlin/live/ditto/demo/kmp/DittoPresence.kt
@@ -1,3 +1,5 @@
+package live.ditto.demo.kmp
+
 import kotlinx.coroutines.flow.Flow
 
 @Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
