+++ composeApp/src/commonMain/kotlin/live/ditto/demo/kmp/KoinModule.kt
@@ -1,3 +1,5 @@
+package live.ditto.demo.kmp
+
 import org.koin.core.context.startKoin
 import org.koin.dsl.KoinAppDeclaration
