+++ composeApp/src/commonMain/kotlin/live/ditto/demo/kmp/Platform.kt
@@ -1,3 +1,5 @@
+package live.ditto.demo.kmp
+
 interface Platform {
     val name: String
 }
