+++ composeApp/src/commonMain/kotlin/live/ditto/demo/kmp/PresenceView.kt
@@ -1,3 +1,5 @@
+package live.ditto.demo.kmp
+
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
