+++ composeApp/src/commonMain/kotlin/live/ditto/demo/kmp/PresenceViewModel.kt
@@ -1,3 +1,5 @@
+package live.ditto.demo.kmp
+
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.SharingStarted
