+++ composeApp/src/iosMain/kotlin/live/ditto/demo/kmp/MainViewController.kt
@@ -1,3 +1,5 @@
+package live.ditto.demo.kmp
+
 import androidx.compose.ui.window.ComposeUIViewController
 
 @Suppress("FunctionName")
