+++ composeApp/src/iosMain/kotlin/live/ditto/demo/kmp/Peer.kt
@@ -1,3 +1,5 @@
+package live.ditto.demo.kmp
+
 import cocoapods.DittoObjC.DITConnection
 import cocoapods.DittoObjC.DITPeer
 import kotlinx.cinterop.ExperimentalForeignApi
