+++ composeApp/src/iosMain/kotlin/live/ditto/demo/kmp/Platform.ios.kt
@@ -1,3 +1,5 @@
+package live.ditto.demo.kmp
+
 import platform.UIKit.UIDevice
 
 class IOSPlatform : Platform {
